Broke out app into individual files
